+++ src/App.tsx
@@ -1,13 +1,18 @@
 import React from "react";
+import { ThemeProvider } from "./contexts/ThemeContext";
 import HoldingCalculator from "./components/HoldingCalculator";
 
 /**
  * App - Root Component
  * 
- * Purpose: Main application entry point
+ * Purpose: Main application entry point with theme provider
  */
 function App() {
-  return <HoldingCalculator />;
+  return (
+    <ThemeProvider>
+      <HoldingCalculator />
+    </ThemeProvider>
+  );
 }
 
 export default App;